refactor component directory
