+++ src/Pages/HomePage.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Feed from "../Components/Feed";
-import Sidebar from "../Components/Sidebar";
+import Sidebar from "../Components/Sidebar/Sidebar";
 
 export default function HomePage() {
   return (
